+++ lib/src/widgets/nes_button.dart
@@ -45,7 +45,10 @@ class NesButton extends StatefulWidget {
     required this.type,
     required this.child,
     this.onPressed,
-  }) : _isDisabled = onPressed == null;
+    this.onPressDown,
+    this.onPressUp,
+  }) : _isDisabled =
+            onPressed == null && onPressDown == null && onPressUp == null;
 
   /// {@macro nes_button}
   ///
@@ -55,12 +58,16 @@ class NesButton extends StatefulWidget {
     required NesButtonType type,
     required NesIconData icon,
     VoidCallback? onPressed,
+    VoidCallback? onPressDown,
+    VoidCallback? onPressUp,
     Size? iconSize,
   }) : this(
           key: key,
           type: type,
           child: NesIcon(iconData: icon, size: iconSize),
           onPressed: onPressed,
+          onPressDown: onPressDown,
+          onPressUp: onPressUp,
         );
 
   /// {@macro nes_button}
@@ -72,6 +79,8 @@ class NesButton extends StatefulWidget {
     required String text,
     double? buttonWidth,
     VoidCallback? onPressed,
+    VoidCallback? onPressDown,
+    VoidCallback? onPressUp,
   }) : this(
           key: key,
           type: type,
@@ -82,6 +91,8 @@ class NesButton extends StatefulWidget {
                 )
               : Text(text),
           onPressed: onPressed,
+          onPressDown: onPressDown,
+          onPressUp: onPressUp,
         );
 
   /// {@macro nes_button}
@@ -93,6 +104,8 @@ class NesButton extends StatefulWidget {
     required NesIconData icon,
     required String text,
     VoidCallback? onPressed,
+    VoidCallback? onPressDown,
+    VoidCallback? onPressUp,
   }) : this(
           key: key,
           type: type,
@@ -105,6 +118,8 @@ class NesButton extends StatefulWidget {
             ],
           ),
           onPressed: onPressed,
+          onPressDown: onPressDown,
+          onPressUp: onPressUp,
         );
 
   /// Determines the colors used to render the button.
@@ -115,6 +130,12 @@ class NesButton extends StatefulWidget {
 
   /// The callback called the button is pressed.
   final VoidCallback? onPressed;
+
+  /// The callback called when the button is pressed down.
+  final VoidCallback? onPressDown;
+
+  /// The callback called when the button is released.
+  final VoidCallback? onPressUp;
 
   /// Whether the button is disabled.
   final bool _isDisabled;
@@ -163,9 +184,11 @@ class _NesButtonState extends State<NesButton> {
       child: GestureDetector(
         onTap: widget.onPressed,
         onTapDown: (_) {
+          widget.onPressDown?.call();
           setState(() => _pressed = true);
         },
         onTapUp: (_) {
+          widget.onPressUp?.call();
           setState(() => _pressed = false);
         },
         onTapCancel: () {