+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - feat: add `buttonsWidth` to `NesConfirmDialog`
  - fix: Fixes for Flutter 3.38.x
  - fix: Snackbar was not proper animating
+ - feat: Enhancing NesButton to add `onPressDown` and `onPressUp`
 
 # 0.27.0
  - feat: adding `buttonWidth` attribute to `NesButton.text` constructor.